Add size_t in the line 14

+++ 0x1E-search_algorithms/0-linear.c
@@ -11,13 +11,13 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	unsigned int i = 0;
+	size_t i = 0;
 
 	if (array == NULL)
 		return (-1);
 	for ( ; i < size; i++)
 	{
-		printf("Value checked array[%d] = [%d]\n", i, array[i]);
+		printf("Value checked array[%ld] = [%d]\n", i, array[i]);
 		if (array[i] == value)
 			return (i);
 	}
